Create a unified IO base class.
This class, IOs, contains public and protected print and prompt methods. This
is done so there is a centralized, non-redundant way to prompt the user and
print verbosity. The public and protected versions exist such that subclasses
may easily remove the public methods.
